+++ src/app/community/page.tsx
@@ -1,6 +1,5 @@
 import Benefits from "@/components/Community/Benefits";
 import CommunityHeader from "@/components/Community/CommunityHeader";
-import MemberBox from "@/components/Community/MemberBox";
 import OurMembers from "@/components/Community/OurMembers";
 
 export default function CommunityPage() {
@@ -9,7 +8,6 @@ export default function CommunityPage() {
       <CommunityHeader />
       <Benefits />
       <OurMembers />
-      <MemberBox />
     </div>
   );
 }
