+++ src/components/Community/Benefits.tsx
@@ -9,6 +9,45 @@ import Expert from "./Logo/Expert";
 import Container from "../Common/Container";
 
 const Benefits = () => {
+  const benefits = [
+    {
+      title: "Connection",
+      description:
+        "Our partners provide valuable connections that can help broaden your professional network. By collaborating with us.",
+      icon: <Connection />,
+    },
+    {
+      title: "Coupon",
+      description:
+        "As part of our collaboration, you will receive exclusive coupons that offer special privileges to join the event or special discount.",
+      icon: <Coupon />,
+    },
+    {
+      title: "Internship",
+      description:
+        "We offer opportunities for internships with leading blockchain companies.",
+      icon: <Internship />,
+    },
+    {
+      title: "Experience",
+      description:
+        "Working with blockchain followers provides a unique experience in Community Engagement.",
+      icon: <Experience />,
+    },
+    {
+      title: "Idea to Impact",
+      description:
+        "Collaborating with us gives you the chance to create and develop new projects.",
+      icon: <Idea />,
+    },
+    {
+      title: "Expert Insight",
+      description:
+        "Benefit from the opportunity to receive valuable advice from industry experts and speakers.",
+      icon: <Expert />,
+    },
+  ];
+
   return (
     <div className="bg-gradient-to-b from-[#0F082C] via-black to-black py-16">
       <Container>
@@ -24,63 +63,17 @@ const Benefits = () => {
           When you join our club, you&#39;ll gain unique experiences, build
           valuable connections, and much more.
         </p>
+
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-10">
-          <div className="flex flex-col items-center">
-            <Connection />
-            <h3 className="mt-4 text-xl font-bold text-white">Connection</h3>
-            <p className="text-gray-300 text-center mt-2">
-              Our partners provide valuable connections that can help broaden
-              your professional network. By collaborating with us.
-            </p>
-          </div>
-
-          <div className="flex flex-col items-center">
-            <Coupon />
-            <h3 className="mt-4 text-xl font-bold text-white">Coupon</h3>
-            <p className="text-gray-300 text-center mt-2">
-              As part of our collaboration, you will receive exclusive coupons
-              that offer special privileges to join the event or special
-              discount.
-            </p>
-          </div>
-
-          <div className="flex flex-col items-center">
-            <Internship />
-            <h3 className="mt-4 text-xl font-bold text-white">Internship</h3>
-            <p className="text-gray-300 text-center mt-2">
-              We offer opportunities for internships with leading blockchain
-              companies.
-            </p>
-          </div>
-
-          <div className="flex flex-col items-center">
-            <Experience />
-            <h3 className="mt-4 text-xl font-bold text-white">Experience</h3>
-            <p className="text-gray-300 text-center mt-2">
-              Working with blockchain followers provides a unique experience in
-              Community Engagement.
-            </p>
-          </div>
-          <div className="flex flex-col items-center">
-            <Idea />
-            <h3 className="mt-4 text-xl font-bold text-white">
-              Idea to Impact
-            </h3>
-            <p className="text-gray-300 text-center mt-2">
-              Collaborating with us gives you the chance to create and develop
-              new projects.
-            </p>
-          </div>
-          <div className="flex flex-col items-center">
-            <Expert />
-            <h3 className="mt-4 text-xl font-bold text-white">
-              Expert Insight
-            </h3>
-            <p className="text-gray-300 text-center mt-2">
-              Benefit from the opportunity to receive valuable advice from
-              industry experts and speakers.
-            </p>
-          </div>
+          {benefits.map(({ title, description, icon }, index) => (
+            <div key={index} className="flex flex-col items-center gap-4">
+              <div className="h-44">{icon}</div>
+              <div className="flex flex-col items-center">
+                <h3 className="mt-4 text-xl font-bold text-white">{title}</h3>
+                <p className="text-gray-300 text-center mt-2">{description}</p>
+              </div>
+            </div>
+          ))}
         </div>
       </Container>
     </div>
